- use the refactored list widgets for the containers tab
